+++ foxes/models/wake_models/wind/bastankhah14.py
@@ -194,7 +194,7 @@ class Bastankhah2014(GaussianWakeModel):
             # calculate sigma:
             # beta = 0.5 * (1 + np.sqrt(1.0 - ct)) / np.sqrt(1.0 - ct)
             a = self.induction.ct2a(ct)
-            beta = (1 - a) / (1 - 2 * a)
+            beta = np.maximum((1 - a) / (1 - 2 * a), 0)
             sigma = k * x + self.sbeta_factor * np.sqrt(beta) * D
             del beta, a