+++ CHANGELOG.md
@@ -798,6 +798,7 @@ This major version introduces the concept of `Engines` which handle the chunking
 
 ## v1.5.2
 
-
+- Tests:
+  - New test `test_set_farm_vars`, checking if correct turbine ordering is maintained
 
 **Full Changelog**: [https://github.com/FraunhoferIWES/foxes/commits/v1.5.2](https://github.com/FraunhoferIWES/foxes/commits/v1.5.2)